+++ test/hdf5lib/TestH5Obasic.java
@@ -23,7 +23,7 @@ public class TestH5Obasic {
     private static long H5la_l1 = -1;
     private static long H5la_dt1 = -1;
     private static long H5la_g1 = -1;
-    int H5fid = -1;
+    long H5fid = -1;
 
     @Before
     public void openH5file()
@@ -49,7 +49,7 @@ public class TestH5Obasic {
 
     @Test(expected = HDF5LibraryException.class)
     public void testH5Oopen_not_exists() throws Throwable {
-        int oid = -1;
+        long oid = -1;
 
         oid = H5.H5Oopen(H5fid, "Never_created", HDF5Constants.H5P_DEFAULT);
 
@@ -58,7 +58,7 @@ public class TestH5Obasic {
 
     @Test
     public void testH5Oget_info_dataset() {
-        int oid = -1;
+        long oid = -1;
         H5O_info_t obj_info = null;
         
         try {
@@ -76,7 +76,7 @@ public class TestH5Obasic {
 
     @Test
     public void testH5Oget_info_hardlink() {
-        int oid = -1;
+        long oid = -1;
         H5O_info_t obj_info = null;
         try {
             oid = H5.H5Oopen(H5fid, "L1", HDF5Constants.H5P_DEFAULT);
@@ -93,7 +93,7 @@ public class TestH5Obasic {
 
     @Test
     public void testH5Oget_info_group() {
-        int oid = -1;
+        long oid = -1;
         H5O_info_t obj_info = null;
         try {
             oid = H5.H5Oopen(H5fid, "G1", HDF5Constants.H5P_DEFAULT);
@@ -110,7 +110,7 @@ public class TestH5Obasic {
 
     @Test
     public void testH5Oget_info_datatype() {
-        int oid = -1;
+        long oid = -1;
         H5O_info_t obj_info = null;
         try {
             oid = H5.H5Oopen(H5fid, "DT1", HDF5Constants.H5P_DEFAULT);
@@ -214,7 +214,7 @@ public class TestH5Obasic {
 
     @Test
     public void testH5Oget_info_by_idx_n0() {
-        int oid = -1;
+        long oid = -1;
         H5O_info_t obj_info = null;
         try {
             oid = H5.H5Oopen(H5fid, "DS1", HDF5Constants.H5P_DEFAULT);
@@ -240,7 +240,7 @@ public class TestH5Obasic {
 
     @Test
     public void testH5Oget_info_by_idx_n3() {
-        int oid = -1;
+        long oid = -1;
         H5O_info_t obj_info = null;
         try {
             oid = H5.H5Oopen(H5fid, "L1", HDF5Constants.H5P_DEFAULT);
@@ -279,7 +279,7 @@ public class TestH5Obasic {
         }
         H5O_iterate_t iter_data = new H5O_iter_data();
         class H5O_iter_callback implements H5O_iterate_cb {
-            public int callback(int group, String name, H5O_info_t info, H5O_iterate_t op_data) {
+            public int callback(long group, String name, H5O_info_t info, H5O_iterate_t op_data) {
                 idata id = new idata(name, info.type);
                 ((H5O_iter_data)op_data).iterdata.add(id);
                 return 0;
@@ -318,7 +318,7 @@ public class TestH5Obasic {
         }
         H5O_iterate_t iter_data = new H5O_iter_data();
         class H5O_iter_callback implements H5O_iterate_cb {
-            public int callback(int group, String name, H5O_info_t info, H5O_iterate_t op_data) {
+            public int callback(long group, String name, H5O_info_t info, H5O_iterate_t op_data) {
                 idata id = new idata(name, info.type);
                 ((H5O_iter_data)op_data).iterdata.add(id);
                 return 0;
